+++ CHANGELOG.md
@@ -10,10 +10,11 @@
 ## Unreleased
 
 * Remove the last traces of jQuery ([PR #2702](https://github.com/alphagov/govuk_publishing_components/pull/2702))
+* Add tracking on the accordion ([#2693](https://github.com/alphagov/govuk_publishing_components/pull/2693))
 
 ## 29.1.0
 
-* Remove "priority breadcrumbs" ([PR #2666](https://github.com/alphagov/govuk_publishing_components/pull/2666))
+* Remove "priority breadcrumbs" ([#2666](https://github.com/alphagov/govuk_publishing_components/pull/2666))
 
 ## 29.0.1